dashboard/requirements.txt
pandas
redshift_connector
python-dotenv
streamlit
plotly